update the dependencies and handling of missing varibles
After some more inspection of the build system it turns out that
now hdf5 is always required, so I removed the useflag and made the
dependency obligarory.

In addition, I changed the handling when the environment variables are not set
i.e. only if python bindings are requested and only for the *DIMS variable
the ebuold will bail out, for the memory size the default fallback "1" will
be used if it is not set.

I also added the sse2 use flag since VNL can profit from it.