sys-cluster/charmdebug: Updated manifest.
The Charm++ developers replaced the tar file on their webpage with a bug fix
release with the same filename and version as before (6.2.0). I updated the
manifest and bumped the ebuild so that users will get the benefit of this new
version. Since the old tar file is not available anymore from the charm
download page, I renamed the ebuild instead of copying it.

Package-Manager: portage-2.2.0_alpha195