sci-chemistry/freeon-1.0 Version bump.
The versioning scheme changed from a date based version to a more common
versioning scheme. Portage considers the latest release, 1.0, to be a
downgrade. I masked the older versions to every so gently suggest an upgrade.
The live version's URI is fixed now, the code moved from savannah to github,
and it is slotted to "live" so that both live and release versions can be
installed simulataneously.

Package-Manager: portage-2.2.1