Add src_unpack to ensure mercurial_src_unpack is called
The inheritance mechanism of ebuilds is such that functions not
overridden are taken from the last package inherited. In this case,
that is vcs-snapshot, which provides a src_unpack that does not clone
the mercurial repository. By explicitly providing src_unpack, we
guarantee that mercurial_src_unpack is called prior to
vcs-snapshot_src_unpack.