sci-biology/plinkseq: fixing issues with bundled dev-libs/protobuf and www-servers/mongoose
The git repo received some commits this year so trying to create an ebuild
for that. It bailed out at more places than the latest release tarball
v0.10. The static Makefile needs more tweaks so that the values from
'pkg-config protobuf' calls get propagated properly. And not to look
for the removed protobuf and mongoose subdirs.

Package-Manager: portage-2.3.2