sci-libs/libmaus2: make configure to enable optional but highly wanted features
The configure by default has a number of features disables, and does not enable
the option. It somehow does not even try to look for headers and libs from say
snappy unless --enable-snappy was passed.

Try to blindly enable most of the features although it is not clear whether say
gnutls should be preferred over nettle, likewise when --enable-sse4 is any better
--enable-sse3 (and should not be enabled concurrently with it).

Package-Manager: Portage-2.3.5, Repoman-2.3.2