sci-biology/trinityrnaseq: version bump (no KEYWORDS set)
The perl include paths are not matching the paths in *.pm files
all *.pm need to be patched so that they could be placed into
site_perl subdirectory.

In brief, the package still does not support installation into
a site-wide directory. users are supposed to executed off the
unpacked tarball directly.

Package-Manager: Portage-2.3.28, Repoman-2.3.9