sci-biology/gclib: removed as useless
https://github.com/gpertea/gclib/issues/2#issuecomment-400125651

There is a bit of confusion here, this "library" is not meant to be
a typical object library (no binary static file is being built) and
certainly not meant to be built and installed globally into some
${prefix}/lib/ with corresponding headers under ${prefix}/include
or so...
It's just a loose collection of reusable source code which I directly
include/source into a few of my bioinformatics projects, at compile
time, as needed. As such there should be no reason to install it
separately, system wide! (and I would advise against it).

Moreover, the Makefile included here was simply meant for some quick
testing during development, not for distribution. I have some simple
scripts included in the actual projects themselves (e.g. gffread)
which are pulling the minimum necessary source code files from the
gclib directory in the process of creating a custom release tarball
for distribution, which can then be found on the official release web
page for each project (e.g. for gffread is
http://ccb.jhu.edu/software/stringtie/dl/gffread-0.9.12.tar.gz). That
"official" source code release tarball already includes the gclib source
files that are needed to build the project -- so I would advise various
distributions to use those releases instead as they should build properly
with the gclib code "frozen" at the time of that particular release.
Since gclib code is in continuous development/tweaking, as the various
projects using it are too, there is no guarantee that the various
automatic github releases are going to be all build compatible all the
time (though I do try to keep the "master" branches in sync and ready
to build, but beyond those "official", manually tested releases, it's
hard to keep multiple projects in sync all the time).