sci-biology/augustus: version bump, fix awful Makefile's
Do not require static libraries. Only bamtools are really needed.
htslib is not needed. For bam2wig samtools-0.2 series is needed.

I decided to blindly install some of the binaries created but not
copied into `./bin/`.

`src/Makefile` will gladly define mysql and sqlite together, probably
a non-sense. We force mysql as we check for it first. Not sure if
sqlite should be preferred.

I documented in the augustus-3.3.2.ebuild other "forgotten" files,
like gbrowse.conf and a template for gff2ps.

I dropped `-O3` but left in `-ggdb` in the `CXXFLAGS`.

https://github.com/Gaius-Augustus/Augustus/issues/14
https://github.com/Gaius-Augustus/Augustus/issues/50
https://github.com/Gaius-Augustus/Augustus/issues/51
https://github.com/Gaius-Augustus/Augustus/issues/52
https://github.com/Gaius-Augustus/Augustus/issues/53

Package-Manager: Portage-2.3.62, Repoman-2.3.12
Signed-off-by: Martin Mokrejs <mmokrejs@fold.natur.cuni.cz>