sci-biology/repeatmasker: rewamp configuration code
The changes are partly based on https://github.com/gentoo/sci/pull/580
from Michael Schubert but it seems we could figure out the paths
to binaries automatically using `which`.

Block between the USE options and allow only a single to be enabled,
that is what configure requires, otherwise it enters a loop.

The hmmer is the default in configure and preferred.

Tested with `USE="hmmer -wublast -phrap -rmblast"` and other
combinations while only one is allowed.

If users have wublast/abblast/rmblast/crossmatch in the PATHS we just
pick it.

`which` returns a '.' (dot) if a binary was not found, we need to zap it
otherwise configure enters a loop.

In the end, one can get:

Add a Search Engine:
   1. CrossMatch: [ Configured ]
   2. RMBlast - NCBI Blast with RepeatMasker extensions: [ Configured ]
   3. WUBlast/ABBlast (required by DupMasker): [ Un-configured ]
   4. HMMER3.1 & DFAM: [ Configured, Default ]

   5. Done

properly written into

/usr/lib64/perl5/vendor_perl/5.28.0/RepeatMaskerConfig.pm

Package-Manager: Portage-2.3.62, Repoman-2.3.12
Signed-off-by: Martin Mokrejs <mmokrejs@fold.natur.cuni.cz>