sci-biology/barnacle: remove package, py2 only, does not compile
This is broken beyond repair, it is not compatible with py3.
I tried to save it with DISTUTILS_USE_SETUPTOOLS="no"
but it does not compile: EOFError: EOF when reading a line

We could patch it, but IMO it is not worth the effort
upstream is dead since 2016.

Package-Manager: Portage-3.0.12, Repoman-3.0.2
Signed-off-by: Andrew Ammerlaan <andrewammerlaan@riseup.net>
Signed-off-by: Horea Christian <chr@chymera.eu>