sci-biology/snphylo: remove pack
- this has missing deps
- has not been fixed since it was added in 2015
- does not work, and will never work without some
crazy workarounds. This package installs a script which looks for
a config file and a directory named scripts in the same directory
where the script is installed. At the moment only the master
script is installed, which of course does not work because
the config files and the .py files in the scripts dir are
missing. We could of course add those, but then they would
have to be in the same dir as the master script. (/opt would
be an option, with a symlink of the executable script in a
directory which is in PATH, though I am not sure if
a symlink will work correctly with the way this package
looks for it's config/scripts files)

This was obviously designed for downloading and placing
as is in some user directory.

Signed-off-by: Andrew Ammerlaan <andrewammerlaan@riseup.net>