sci-libs/fox: add inherit fortran-2
Currently, sci-libs/fox does not require a fortran compiler
to be installed. If sci-libs/fox is emerged without the fortran
compiler, the configure phase will check for a fortran compiler, and the
configure phase will fail. For example, the hardened profile disables
the fortran USE flag in sys-devel/gcc, so sci-libs/fox will fail.
With this commit, sci-libs/fox will require a fortran compiler
to be installed.
Without this commit, sci-libs/fox will fail on any system which does
not have the fortran compiler, which by default is all hardened profiles.
This commit was tested in a docker image with dev-util/ebuildtester.
This commit was written, tested, and submitted by Lucas Mitrak.

Closes: https://bugs.gentoo.org/757771
Signed-off-by: Lucas Mitrak <lucas@lucasmitrak.com>
Closes: https://github.com/gentoo/sci/pull/1083
Signed-off-by: Andrew Ammerlaan <andrewammerlaan@gentoo.org>