media-libs/libgfx: add test, dependencies, patches
* Add test, dependencies, and new patches to ebuild
* Add cstring to needed files due to newer version of gcc
* Change png jmpbuf to use the newer safer method due to GLSA 200408-03

Currently, media-libs/libgfx will not compile due to a vulnerability in
libpng which has since been patched [1]. Therefore, the patch
libPNG-1.2.5 updates the code to libPNG's newer and safer method. This
also fixes bug https://bugs.gentoo.org/756061.
However, even with this patch the package will
not compile due to outdated C++, so the patch gcc-4.3 updates two files
to include cstring, which is neccessary for memcpy. These patches will
allow the package to compile. Without these patches, the package will
not compile.
In addition, the ebuild was updated to include these patches, as well as
add the dependencies listed on upstream's homepage [2]. Without these
dependencies, the package will not compile. Finally, a test IUSE flag as
well as the src_test() function were added. Due to the package using
outdated FLTK code, some of the tests had to be disabled using a sed
command.
This commit was tested in a docker image with dev-util/ebuildtester.
This commit was written, tested, and submitted by Lucas Mitrak.

[1] https://security.gentoo.org/glsa/200408-03
[2] http://mgarland.org/software/libgfx.html
Closes: https://bugs.gentoo.org/756061
Signed-off-by: Lucas Mitrak <lucas@lucasmitrak.com>
Signed-off-by: Andrew Ammerlaan <andrewammerlaan@gentoo.org>