sci-biology/dcm2niix: add patch
* Disable finding git in cmake
* Add patch to ebuild

Currently, sci-biology/dcm2niix will not compile if git is not found.
The patch dcm2niix-disable_find_git.patch comments out the lines which
attempt to find git and require it to be found. With this patch, the
package will compile correctly. Without this patch, cmake will throw
an error if git is not found.
This commit was tested in a docker image with dev-util/ebuildtester.
This commit was written, tested, and submitted by Lucas Mitrak.

Closes: https://bugs.gentoo.org/755746
Signed-off-by: Lucas Mitrak <lucas@lucasmitrak.com>
Closes: https://github.com/gentoo/sci/pull/1091
Signed-off-by: Andrew Ammerlaan <andrewammerlaan@gentoo.org>