sci-biology/dcm2niix: add USE flags, dependencies, src_configure
* Add static system-jpeg +jpeg-ls jpeg2k use flags and implement them

Currently, sci-biology/dcm2niix does not take advantage of many
of the package's potential features. These potential features include
a static runtime, using the system-wide media-libs/libjpeg-turbo
instead of bundled compact NanoJPEG decoder, enabling JPEG-LS using
the bundled CharLS library, and enabling JPEG2000 support. These are
important, especially since the static runtime is enabled by default
and JPEG-LS can be enabled without any downside. This commit adds
these USE flags and their implementation in the dependencies and in
src_configure. The naming of the use flags follow the same naming
convention found elsewhere. For example, system-jpeg is the same
as thunderbird, firefox, and seamonkey. In addition, jpeg-ls is the
only possible name for the JPEG-LS transfer syntaxes. These two
package-specific USE flag names were added to metadata.xml.
With this commit, the use flags will be added to the package.
Without this commit, the package will not have these potential
features.
This commit was tested in a docker image with dev-util/ebuildtester.
This commit was written, tested, and submitted by Lucas Mitrak.
Signed-off-by: Lucas Mitrak <lucas@lucasmitrak.com>