scripts/setup-and-run-repoman.sh: use 'sort -V' to get latest tag
Without the -V flag to enable version sort, 'sort' can falsely identify
a version string like 3.0.9 to be greater than 3.0.10.  This can be
shown by running 'git tag | grep portage | sort -u' in the Portage Git
tree.  This issue is not so prominent for repoman yet because the latest
version of repoman as of now is 3.0.3, but it will surface when repoman
3.0.10 is released.

Signed-off-by: Yuan Liao <liaoyuan@gmail.com>
Closes: https://github.com/gentoo/sci/pull/1094
Signed-off-by: Andrew Ammerlaan <andrewammerlaan@gentoo.org>