sci-biology/SEECER: EAPI bump, patch, remove files, fix build system
* EAPI bump
* Remove "rename_jellyfish_binary.patch" from ebuild which does nothing
* Add a patch that increases the max sequence length
* Add a patch that removes CXXFLAGS, CFLAGS, LDFLAGS
* Add tc-check-openmp so the ebuild fails without OpenMP support
* Remove copying a new replace_ids.cc file which is now a patch
* Delete all object files
* Delete packaged seqan
* Add eautoreconf to recreate the build system
* Add system seqan include path and CXXFLAGS to econf

Currently, sci-biology/SEECER will not compile due to object files
remaining in the package. This was fixed by deleting all object files
and so forcing them to be rebuilt. The packaged seqan was deleted as
well so the system's seqan is used. In order to achieve this, the
system's seqan include path was passed into econf as
"/usr/include/seqan". The user's CXXFLAGS were passed into econf also.
The patch SEECER-remove-flags.patch removes any other CXXFLAGS, CFLAGS,
and LDFLAGS present in configure.ac and Makefile.am.
Since configure.ac and Makefile.am were updated, the build system was
recreated using eautoreconf -i to account for this. The -i was needed
because the "compile" file is missing, so it had to be added. The patch
SEECER-increase-max-sequence-length.patch increases the max sequence
length from 256 to 10000. This patch was made from an updated
replace_ids.cc downloaded from an archived upstream mail list
that is no longer available online [1]. The previous revision ebuild
included this same replace_ids.cc file; however, despite it appearing
as a text file, it was actually a gzipped file with an incorrect
extension. In addition, this gzipped file was copied into the src/
directory when the replace_ids.cc file is in the pipeline/ directory.
The patch file "rename_jellyfish_binary.patch" also did nothing
but modify a comment so it was removed as well. In addition,
a tc-check-openmp was added so the ebuild fails without OpenMP support.
OpenMP cannot be added as a use flag or a configure enable flag because
gomp is always enabled and the source code assumes the system has OpenMP
installed. Finally, the EAPI was bumped.

[1] URL Split into three parts for the sake of the commit message:
https://web.archive.org/web/20150911132231/
http://seecer-rna-read-error-correction-mailing-list.21961.x6.nabble.com
/Segmentation-fault-in-step-4-td41.html

Closes: https://bugs.gentoo.org/755776
Package-Manager: Portage-3.0.20, Repoman-3.0.2
Signed-off-by: Lucas Mitrak <lucas@lucasmitrak.com>
Closes: https://github.com/gentoo/sci/pull/1105
Signed-off-by: Andrew Ammerlaan <andrewammerlaan@gentoo.org>