sci-misc/europa-bin: add new package
The official electron client for jupyterlab (jupyterlab_app) is inactive and
freezes upon launch.
This is one of the few alternatives that: has releases, is up to date, and
actually works.

It works rather nice and avoids having to run jupyterlab in the browser.
Thanks to @suyashmahar !

https://github.com/jupyterlab/jupyterlab_app/issues/186

Package-Manager: Portage-3.0.22, Repoman-3.0.3
Signed-off-by: Andrew Ammerlaan <andrewammerlaan@gentoo.org>