sci-misc/irstlm: remove AM_CXXFLAGS, add doc/static-libs use flags
* EAPI bump
* Add doc and static-libs use flags
* Add app-text/texlive[extra] as BDEPEND for doc
* Add patch that removes lib and lib64 linking
* Add patch that makes irstlm-manual.pdf obey --docdir
* Remove AM_CXXFLAGS that breaks the package or should not be there
* Copy doc/RELEASE to RELEASE.tex for doc
* Add static-libs and doc to the configure using the econf variable

Currently, sci-misc/irstlm will not compile due to bad AM_CXXFLAGS.
A sed command was added which removes these unwanted flags. This was
done in the same way to a PR upstream [1]. In addition, a patch was
added which removes a Makefile.am lib->lib64 linking operation. This
was also done in the same way as the upstream PR [1]. Another patch was
added which removes doc/irstlm-manual.pdf from install in --prefix to
instead obey --docdir. Furthermore, the doc USE flag was added for this
new documentation. To create this documentation, a simple copy command
of doc/RELEASE to RELEASE.tex is needed for it to make correctly. Also,
a BDEPEND of app-text/texlive[extra] was added if the doc USE is true.
A static-libs USE flag was added to include the static library
libirstlm.a. Both USE flags are configured in the econf variable.
Finally, the EAPI was bumped.

[1] https://github.com/irstlm-team/irstlm/pull/12

Closes: https://bugs.gentoo.org/755473
Package-Manager: Portage-3.0.20, Repoman-3.0.3
Signed-off-by: Lucas Mitrak <lucas@lucasmitrak.com>
Closes: https://github.com/gentoo/sci/pull/1112
Signed-off-by: Andrew Ammerlaan <andrewammerlaan@gentoo.org>