sci-mathematics/acl2: Require >=dev-lisp/sbcl-1.5.2 in BDEPEND
* Require >=dev-lisp/sbcl-1.5.2 due to addition of "--tls-limit 16384"

Currently, sci-mathematics/acl2-8.4 will not compile with the current
stable dev-lisp/sbcl-1.4.9 due to the addition of "--tls-limit 16384"
to sbcl within the created file saved_acl2. This addition is documented
in the acl-8.4 release notes [1] as "Increased the number of special
variables that can be created". Within sbcl, this new feature was added
in version 1.5.2. The news for that version [2] reads "enhancement:
command-line option '--tls-limit' can be used". Therefore attempting
to use the argument "--tls-limit 16384" with sbcl versions prior to
1.5.2 causes it to fail. Thus a minimum version requirement is needed
for acl2-8.4.

[1] https://www.cs.utexas.edu/users/moore/acl2/v8-4/combined-manual/
index.html?topic=ACL2____NOTE-8-4
[2] http://www.sbcl.org/all-news.html

Closes: https://github.com/gentoo/sci/issues/1113
Package-Manager: Portage-3.0.20, Repoman-3.0.3
Signed-off-by: Lucas Mitrak <lucas@lucasmitrak.com>
Signed-off-by: Andrew Ammerlaan <andrewammerlaan@gentoo.org>