sci-mathematics/scilab-bin: add version 6.1.1
I officially give up on making a non-bin ebuild. There are several problems:

1. The java dependencies are extremely difficult because of packaging systems
that portage does not support (gralde, maven etc). We can hack around this
as we did in dev-java/lucene::sci, but that is *a lot* of work, for a lot of
dependencies that has to be redone on each version bump.
We can solve this problem by getting the pre-compiled java dependencies from
https://github.com/scilab/scilab-prerequirements. However....

2. Scilab requires versions of libraries that we do not package any more in
::gentoo (e.g. hdf5). And adding and maintaining ebuilds for those old libraries
is too much work.

3. If we take *only* the java stuff from scilab/scilab-prerequirements, build
fails (likely incompatibility with certain versions of some non-java libraries,
see 2.)

4. If we take *all* of scilab/scilab-prerequirements build still fails (with
a different error this time).

5. The amount of scilab dependencies that are being last-rited and I therefore
have to move to ::science to keep the dependencies resolving is increasing.
Scilab is the number 1 cause of ::science breaking at the moment.

Long story short, please enjoy this pre-compiled binary ebuild that actually
works and will save us all a lot of headache.

Closes: https://bugs.gentoo.org/237572
Closes: https://bugs.gentoo.org/691272
Closes: https://github.com/gentoo/sci/issues/879
Closes: https://github.com/gentoo/sci/issues/774
Package-Manager: Portage-3.0.30, Repoman-3.0.3
Signed-off-by: Andrew Ammerlaan <andrewammerlaan@gentoo.org>